perf: Speed up file_tree lookups by ~7x
Switched from doing a linear search over the children vector
to a map lookup.
This lead to a ~7x speedup, from 30s for reading a ~115MiB file to ~4s.
I assumed the speed gain would be less severe, since a linear
search can be faster for small N.

To further optimize one could implement a threshold for using or
even storing the map, where a linear search should be faster.
We could also stop storing the children vector at all.